Combine two system specs for RBD
